pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docassist"
version = "1.0.0"
description = "IRS Form Field Extraction System using Vision-Language Models"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "Pillow>=10.0.0",
    "PyMuPDF>=1.23.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "setfit>=1.0.0",
    "peft>=0.10.0",
    "accelerate>=0.30.0",
    "bitsandbytes>=0.43.0",
    "pyyaml>=6.0.0",
    "pytest>=7.0.0"
]

[project.optional-dependencies]
dev = ["pytest", "black", "ruff", "mypy"]

[project.scripts]
docassist = "run:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
